requirements.txt
rank-bm25>=0.2.2
Pillow>=10.0
requests>=2.31
openai>=1.0
tqdm>=4.65
numpy>=1.24
